V4 stats fixer

+++ code/src/main/java/it/uniroma2/pmcsn/parks/model/job/RiderGroup.java
@@ -4,19 +4,19 @@ import it.uniroma2.pmcsn.parks.model.stats.GroupStats;
 
 public class RiderGroup {
 
-    private int groupId;
+    private Long groupId;
     private int groupSize;
     private GroupPriority priority;
     private GroupStats stats;
 
-    public RiderGroup(int groupId, int groupSize, GroupPriority groupPriority, double systemEntranceTime) {
+    public RiderGroup(Long groupId, int groupSize, GroupPriority groupPriority, double systemEntranceTime) {
         this.groupId = groupId;
         this.groupSize = groupSize;
         this.priority = groupPriority;
         this.stats = new GroupStats(systemEntranceTime);
     }
 
-    public int getGroupId() {
+    public Long getGroupId() {
         return this.groupId;
     }
